requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.5.0
django-environ==0.9.0
django-js-asset==2.0.0
djangorestframework==3.14.0
gunicorn==20.0.4
idna==3.4
Pillow==9.3.0
psycopg2-binary==2.9.9
pytz==2022.6
requests==2.31.0
sqlparse==0.4.3
texttable==1.0.0
urllib3==2.0.7
httpx==0.26.0